Hack-A-War-main/HACKAWAR-BACKREP-main/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
boto3>=1.28.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
tenacity>=8.0.0
aiohttp>=3.9.0
requests>=2.31.0
